+++ compiler/syntax/src/res_parens.ml
@@ -162,8 +162,7 @@ let rhs_binary_expr_operand parent_operator rhs =
           };
         args = [(_, _left); (_, _right)];
       }
-    when ParsetreeViewer.is_binary_operator operator
-         && not (operator_loc.loc_ghost && operator = "++") ->
+    when ParsetreeViewer.not_ghost_operator operator operator_loc ->
     let prec_parent = ParsetreeViewer.operator_precedence parent_operator in
     let prec_child = ParsetreeViewer.operator_precedence operator in
     prec_parent == prec_child
@@ -180,8 +179,7 @@ let flatten_operand_rhs parent_operator rhs =
           };
         args = [(_, _left); (_, _right)];
       }
-    when ParsetreeViewer.is_binary_operator operator
-         && not (operator_loc.loc_ghost && operator = "++") ->
+    when ParsetreeViewer.not_ghost_operator operator operator_loc ->
     let prec_parent = ParsetreeViewer.operator_precedence parent_operator in
     let prec_child = ParsetreeViewer.operator_precedence operator in
     prec_parent >= prec_child || rhs.pexp_attributes <> []