+++ compiler/syntax/src/res_token.ml
@@ -78,6 +78,7 @@ type t =
   | Land
   | Lor
   | Band (* Bitwise and: & *)
+  | Caret
   | BangEqual
   | BangEqualEqual
   | LessEqual
@@ -102,14 +103,15 @@ let precedence = function
   | HashEqual | ColonEqual -> 1
   | Lor -> 2
   | Land -> 3
+  | Caret -> 4
   | Equal | EqualEqual | EqualEqualEqual | LessThan | GreaterThan | BangEqual
   | BangEqualEqual | LessEqual | GreaterEqual | BarGreater ->
-    4
-  | Plus | PlusDot | Minus | MinusDot | PlusPlus -> 5
-  | Asterisk | AsteriskDot | Forwardslash | ForwardslashDot | Percent -> 6
-  | Exponentiation -> 7
-  | MinusGreater -> 8
-  | Dot -> 9
+    5
+  | Plus | PlusDot | Minus | MinusDot | PlusPlus -> 6
+  | Asterisk | AsteriskDot | Forwardslash | ForwardslashDot | Percent -> 7
+  | Exponentiation -> 8
+  | MinusGreater -> 9
+  | Dot -> 10
   | _ -> 0
 
 let to_string = function
@@ -189,6 +191,7 @@ let to_string = function
   | Of -> "of"
   | Lor -> "||"
   | Band -> "&"
+  | Caret -> "^"
   | Land -> "&&"
   | BangEqual -> "!="
   | BangEqualEqual -> "!=="