+++ tests/tests/src/belt_int_test.mjs
@@ -36,6 +36,7 @@ Mocha.describe("Belt_int_test", () => {
     Test_utils.eq("File \"belt_int_test.res\", line 41, characters 7-14", 6, 6);
     Test_utils.eq("File \"belt_int_test.res\", line 42, characters 7-14", 0, 0);
     Test_utils.eq("File \"belt_int_test.res\", line 43, characters 7-14", 0, 0);
+    Test_utils.eq("File \"belt_int_test.res\", line 44, characters 7-14", 1, 1);
   });
 });